pyproject.toml
# Largely inspired from (pointing to precise version)
# https://github.com/thejohnfreeman/project-template-python/tree/6d04c7b3b00460bb7473246096c52cc22d403226

# black support portion inspired from (pointing to precise version)
# https://github.com/python/black/blob/4a953b7241ce5f8bcac985fa33fdf3af4f42c0de/pyproject.toml

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to the project.
  # serial_asyncio with a correction (need to declare bug and refer to it here)
  | serial_asyncio
)/
'''

[tool.isort]
# Setup of for black compatibility
# (see https://black.readthedocs.io/en/stable/the_black_code_style.html#how-black-wraps-lines)
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.poetry]
name = "pyteleinfo"
version = "0.0.1"
description = ""
authors = ["Emmanuel Sciara <emmanuel.sciara@gmail.com>"]
packages = [
    { include = "teleinfo" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    'Intended Audience :: End Users/Desktop',
    'Natural Language :: English',
    'Operating System :: POSIX',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: MacOS :: MacOS X',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Topic :: Communications',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Terminals :: Serial',
    'License :: OSI Approved :: BSD License',
]

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "^3.4"
cleo = "^0.7.4"

[tool.poetry.dev-dependencies]
pytest = "^4.5"
pytest-html = "^1.20"
pytest-asyncio = "^0.10.0"
PyHamcrest = "^1.9"
"path.py" = "^12.0; python <3.6"
bump2version = "^0.5.10"
invoke = "^1.2"
mypy = "^0.701.0"
pydocstyle = "^3.0"
pylint = "^2.3"
pytest-cov = "^2.7"
sphinx = "^2.0"
sphinx-autobuild = "^0.7.1"
sphinx_rtd_theme = "^0.4.3"
toml = "^0.10.0"
yapf = "^0.27.0"
sphinx_bootstrap_theme = "^0.7.1"
black = {version = "^18.3-alpha.0", python = "^3.6"}
nest_asyncio = "^1.0"
pre-commit = "^1.16"
isort = "^4.3"
flake8 = "^3.7"
flake8-bugbear = "^19.3"
tox = "^3.12"
tox-pyenv = "^1.1"

[tool.poetry.scripts]
teleinfo = "teleinfo.console:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
